+++ packages/eslint-plugin/src/index.cjs
@@ -12,6 +12,7 @@ module.exports = {
   rules: {
     "import-playwright-a11y": require("./rules/import-playwright-a11y.cjs"),
     "no-shadow-native": require("./rules/no-shadow-native-events.cjs"),
+    "no-shadow-template-ref": require("./rules/no-shadow-template-ref.cjs"),
     "require-root-class": require("./rules/require-root-class.cjs"),
   },
 };