+++ packages/headless/src/composables/tabs/TestTabs.ct.tsx
@@ -2,7 +2,6 @@ import { test } from "@playwright/experimental-ct-vue";
 import TestTabs from "./TestTabs.vue";
 import { tabsTesting } from "./createTabs.testing";
 
-// eslint-disable-next-line playwright/expect-expect
 test("tabs", async ({ mount, page }) => {
   const component = await mount(<TestTabs />);