+++ packages/sit-onyx/src/components/OnyxNotificationCard/OnyxNotificationCard.ct.tsx
@@ -1,7 +1,8 @@
 import { DENSITIES } from "../../composables/density";
 import { expect, test } from "../../playwright/a11y";
-import { executeMatrixScreenshotTest } from "../../playwright/screenshots";
+import { executeMatrixScreenshotTest, mockPlaywrightIcon } from "../../playwright/screenshots";
 import OnyxButton from "../OnyxButton/OnyxButton.vue";
+import OnyxMenuItem from "../OnyxNavBar/modules/OnyxMenuItem/OnyxMenuItem.vue";
 import OnyxNotificationCard from "./OnyxNotificationCard.vue";
 
 const MOCK_DATE = new Date(2025, 3, 2, 13, 45);
@@ -19,7 +20,7 @@ test.describe("Screenshot tests", () => {
   executeMatrixScreenshotTest({
     name: "Notification card",
     columns: DENSITIES,
-    rows: ["default", "unread", "actions", "separator"],
+    rows: ["default", "unread", "actions", "separator", "icon"],
     component: (column, row) => {
       const card = (
         <OnyxNotificationCard
@@ -28,6 +29,7 @@ test.describe("Screenshot tests", () => {
           unread={row === "unread"}
           density={column}
           style={{ width: "24rem" }}
+          icon={row === "icon" ? mockPlaywrightIcon : undefined}
         >
           Lorem ipsum dolor sit amet consectetur. Dui purus quisque est varius vulputate.
           {row === "actions" && (
@@ -45,6 +47,40 @@ test.describe("Screenshot tests", () => {
           {row === "separator" && card}
         </div>
       );
+    },
+  });
+});
+
+test.describe("Screenshot tests (header actions)", () => {
+  executeMatrixScreenshotTest({
+    name: "Notification card (header actions)",
+    columns: ["default", "unread"],
+    rows: ["default", "hover", "open"],
+    component: (column) => {
+      return (
+        <OnyxNotificationCard
+          headline="Example notification"
+          createdAt={MOCK_DATE}
+          unread={column === "unread"}
+          style={{ width: "24rem" }}
+        >
+          Lorem ipsum dolor sit amet consectetur. Dui purus quisque est varius vulputate.
+          <template v-slot:headerActions>
+            <OnyxMenuItem>Action 1</OnyxMenuItem>
+            <OnyxMenuItem>Action 2</OnyxMenuItem>
+          </template>
+        </OnyxNotificationCard>
+      );
+    },
+    hooks: {
+      beforeEach: async (component, page, column, row) => {
+        if (row === "hover" || row == "open") {
+          await component.hover();
+        }
+        if (row == "open") {
+          await component.getByLabel("Toggle actions").click();
+        }
+      },
     },
   });
 });
@@ -84,4 +120,51 @@ test("should display elapsed time correctly", async ({ page, mount }) => {
     // ASSERT
     await expect(component).toContainText(label);
   }
+});
+
+test("should show header actions on hover", async ({ mount, page }) => {
+  // ARRANGE
+  const component = await mount(
+    <OnyxNotificationCard
+      headline="Example headline"
+      createdAt={MOCK_DATE}
+      style={{ margin: "3rem" }}
+    >
+      <template v-slot:headerActions>
+        <OnyxMenuItem>Action 1</OnyxMenuItem>
+        <OnyxMenuItem>Action 2</OnyxMenuItem>
+      </template>
+    </OnyxNotificationCard>,
+  );
+
+  const flyoutTrigger = component.getByLabel("Toggle actions");
+  const flyoutMenu = component.getByLabel("More actions");
+
+  // ACT
+  await component.hover();
+
+  // ASSERT
+  await expect(flyoutTrigger, "should show flyout system button on hover").toBeVisible();
+  await expect(flyoutMenu, "should not show menu on hover").toBeHidden();
+
+  // ACT
+  await flyoutTrigger.click();
+
+  // ASSERT
+  await expect(flyoutMenu.getByRole("menuitem", { name: "Action 1" })).toBeVisible();
+  await expect(flyoutMenu.getByRole("menuitem", { name: "Action 2" })).toBeVisible();
+
+  // ACT (reset mouse)
+  await page.getByRole("document").click();
+  await page.getByRole("document").hover({ position: { x: 0, y: 0 } });
+
+  // ASSERT
+  await expect(flyoutMenu).toBeHidden();
+
+  // ACT
+  await component.press("Tab");
+
+  // ASSERT
+  await expect(flyoutTrigger, "should focus trigger with keyboard").toBeFocused();
+  await expect(flyoutMenu, "should show menu on keyboard focus").toBeVisible();
 });