+++ packages/sit-onyx/src/components/OnyxNotificationCard/OnyxNotificationCard.vue
@@ -1,10 +1,14 @@
 <script lang="ts" setup>
+import moreVertical from "@sit-onyx/icons/more-vertical.svg?raw";
 import { toRef } from "vue";
 import { useDensity } from "../../composables/density";
 import { useRelativeTimeFormat } from "../../composables/useRelativeTimeFormat";
 import { injectI18n } from "../../i18n";
 import OnyxBadge from "../OnyxBadge/OnyxBadge.vue";
 import OnyxHeadline from "../OnyxHeadline/OnyxHeadline.vue";
+import OnyxIcon from "../OnyxIcon/OnyxIcon.vue";
+import OnyxFlyoutMenu from "../OnyxNavBar/modules/OnyxFlyoutMenu/OnyxFlyoutMenu.vue";
+import OnyxSystemButton from "../OnyxSystemButton/OnyxSystemButton.vue";
 import type { OnyxNotificationCardProps } from "./types";
 
 const props = defineProps<OnyxNotificationCardProps>();
@@ -15,13 +19,19 @@ const slots = defineSlots<{
    */
   default(): unknown;
   /**
-   * Optional custom actions/buttons.
+   * Optional custom footer actions/buttons.
    */
   actions?(): unknown;
+  /**
+   * Optional custom header actions to display inside a flyout menu.
+   * Will only be shown when hovering the notification card or focussing via keyboard.
+   * You must only put [OnyxMenuItem](http://localhost:6006/?path=/docs/navigation-modules-menuitem--docs) components here.
+   */
+  headerActions?: unknown;
 }>();
 
 const { densityClass } = useDensity(props);
-const { d } = injectI18n();
+const { d, t } = injectI18n();
 const { timeAgo } = useRelativeTimeFormat({
   time: toRef(props, "createdAt"),
   options: { numeric: "auto" },
@@ -31,10 +41,35 @@ const { timeAgo } = useRelativeTimeFormat({
 <template>
   <div :class="['onyx-component', 'onyx-notification-card', densityClass]">
     <div class="onyx-notification-card__content">
-      <div class="onyx-notification-card__header">
-        <div class="onyx-notification-card__headline">
-          <OnyxHeadline is="h3">{{ props.headline }}</OnyxHeadline>
-          <OnyxBadge v-if="props.unread" dot />
+      <div>
+        <div class="onyx-notification-card__header">
+          <div class="onyx-notification-card__header-container">
+            <OnyxIcon v-if="props.icon" class="onyx-notification-card__icon" :icon="props.icon" />
+            <OnyxHeadline is="h3">{{ props.headline }}</OnyxHeadline>
+          </div>
+
+          <div class="onyx-notification-card__header-container">
+            <OnyxBadge v-if="props.unread" dot />
+
+            <OnyxFlyoutMenu
+              v-if="!!slots.headerActions"
+              class="onyx-notification-card__more-actions"
+              :label="t('notificationCard.moreActions')"
+              trigger="click"
+            >
+              <template #button="{ trigger }">
+                <OnyxSystemButton
+                  v-bind="trigger"
+                  :label="t('notificationCard.toggleActions')"
+                  :icon="moreVertical"
+                />
+              </template>
+
+              <template #options>
+                <slot name="headerActions"></slot>
+              </template>
+            </OnyxFlyoutMenu>
+          </div>
         </div>
 
         <div class="onyx-notification-card__created-at onyx-text--small">
@@ -56,6 +91,7 @@ const { timeAgo } = useRelativeTimeFormat({
 
 <style lang="scss">
 @use "../../styles/mixins/layers.scss";
+@use "../../styles/mixins/visibility.scss";
 
 .onyx-notification-card {
   @include layers.component() {
@@ -64,6 +100,11 @@ const { timeAgo } = useRelativeTimeFormat({
     padding: var(--onyx-notification-card-padding);
     background-color: var(--onyx-color-base-background-blank);
     max-width: 100%;
+
+    &:hover,
+    &:focus-within {
+      background-color: var(--onyx-color-base-background-tinted);
+    }
 
     &__content {
       display: flex;
@@ -82,8 +123,37 @@ const { timeAgo } = useRelativeTimeFormat({
 
     &__header {
       display: flex;
-      flex-direction: column;
-      gap: var(--onyx-density-2xs);
+      align-items: center;
+      justify-content: space-between;
+      gap: var(--onyx-density-md);
+      margin-bottom: var(--onyx-density-2xs);
+    }
+
+    &__header-container {
+      display: flex;
+      gap: var(--onyx-density-xs);
+      align-items: center;
+      flex-wrap: wrap;
+    }
+
+    &__icon {
+      color: var(--onyx-color-text-icons-neutral-medium);
+    }
+
+    &__more-actions {
+      .onyx-flyout-menu__list {
+        right: 0;
+      }
+    }
+
+    // for devices that support hover, we are only showing the more actions on hover
+    // for all other devices (like mobile devices), they are always shown
+    @media (hover: hover) {
+      &:not(&:hover, &:focus-within) {
+        .onyx-notification-card__more-actions {
+          @include visibility.visually-hidden();
+        }
+      }
     }
 
     &__description {
@@ -105,13 +175,6 @@ const { timeAgo } = useRelativeTimeFormat({
       gap: var(--onyx-density-md);
       flex-wrap: wrap;
       color: var(--onyx-color-text-icons-neutral-soft);
-    }
-
-    &__headline {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      gap: var(--onyx-density-md);
     }
   }
 }